Update Function Find Factorial.cpp

+++ Day11/Function Find Factorial.cpp
@@ -1,7 +1,7 @@
 #include<iostream>
 using namespace std;
-long long factorial(int n){
-    long long fact=1;
+int factorial(int n){
+    int fact=1;
     for(int i=1;i<=n;i++) fact=fact*i;
     return fact;
 }
